requirements.txt
asgiref==3.7.2
Django==3.2.19
django-cors-headers==4.0.0
django-rest-auth==0.9.5
djangorestframework==3.14.0
pytz==2023.3
six==1.16.0
sqlparse==0.4.4
whitenoise==6.4.0
